+++ app/src/main/java/vn/asiantech/way/extension/Views.kt
@@ -1,12 +1,14 @@
 package vn.asiantech.way.extension
 
+import android.app.Dialog
 import android.content.Context
+import android.util.DisplayMetrics
 import android.view.View
 import android.view.ViewManager
+import android.view.WindowManager
 import android.view.inputmethod.InputMethodManager
 import com.hypertrack.lib.internal.consumer.view.RippleView
 import org.jetbrains.anko.custom.ankoView
-import vn.asiantech.way.ui.custom.BottomButtonCard
 
 /**
  * Created by haingoq on 12/10/2017.
@@ -22,6 +24,13 @@ internal fun ViewManager.rippleView(init: RippleView.() -> Unit): RippleView {
             { RippleView(it, null) }, 0, init)
 }
 
-internal fun ViewManager.bottomCard(init: BottomButtonCard.() -> Unit): BottomButtonCard {
-    return ankoView({ BottomButtonCard(it) }, 0, init)
+internal fun Dialog.setDialogScreenSize(context: Context) {
+    val displayMetrics: DisplayMetrics = context.resources.displayMetrics
+    val width = displayMetrics.widthPixels
+    val height = displayMetrics.heightPixels
+    val windowParams = WindowManager.LayoutParams()
+    windowParams.copyFrom(this.window.attributes)
+    windowParams.height = height - height / 5
+    windowParams.width = width - 80
+    this.window.attributes = windowParams
 }
