+++ app/src/main/java/vn/asiantech/way/ui/share/ShareActivityUI.kt
@@ -7,8 +7,9 @@ import android.widget.ImageView
 import android.widget.TextView
 import org.jetbrains.anko.*
 import vn.asiantech.way.R
-import vn.asiantech.way.extension.bottomCard
 import vn.asiantech.way.extension.rippleView
+import vn.asiantech.way.ui.custom.bottomCard
+
 
 /**
  * Copyright © 2017 Asian Tech Co., Ltd.
